chore(starknet_infra_utils): unify test util mods
commit-id:3470ac8e
