+++ crates/starknet_infra_utils/src/lib.rs
@@ -4,8 +4,6 @@ pub mod path;
 pub mod run_until;
 pub mod tasks;
 #[cfg(any(feature = "testing", test))]
-pub mod test_identifiers;
-#[cfg(any(feature = "testing", test))]
 pub mod test_utils;
 pub mod tracing;
 pub mod type_name;
