--- crates/starknet_infra_utils/src/test_identifiers.rs
@@ -1,18 +1,0 @@
-#[derive(Debug, Copy, Clone)]
-pub enum TestIdentifier {
-    EndToEndIntegrationTest,
-    EndToEndFlowTest,
-    MempoolSendsTxToOtherPeerTest,
-    MempoolReceivesTxFromOtherPeerTest,
-}
-
-impl From<TestIdentifier> for u16 {
-    fn from(variant: TestIdentifier) -> Self {
-        match variant {
-            TestIdentifier::EndToEndIntegrationTest => 0,
-            TestIdentifier::EndToEndFlowTest => 1,
-            TestIdentifier::MempoolSendsTxToOtherPeerTest => 2,
-            TestIdentifier::MempoolReceivesTxFromOtherPeerTest => 3,
-        }
-    }
-}
