+++ crates/starknet_infra_utils/src/test_utils.rs
@@ -16,6 +16,25 @@ const _: () = {
     );
 };
 
+#[derive(Debug, Copy, Clone)]
+pub enum TestIdentifier {
+    EndToEndIntegrationTest,
+    EndToEndFlowTest,
+    MempoolSendsTxToOtherPeerTest,
+    MempoolReceivesTxFromOtherPeerTest,
+}
+
+impl From<TestIdentifier> for u16 {
+    fn from(variant: TestIdentifier) -> Self {
+        match variant {
+            TestIdentifier::EndToEndIntegrationTest => 0,
+            TestIdentifier::EndToEndFlowTest => 1,
+            TestIdentifier::MempoolSendsTxToOtherPeerTest => 2,
+            TestIdentifier::MempoolReceivesTxFromOtherPeerTest => 3,
+        }
+    }
+}
+
 #[derive(Debug)]
 pub struct AvailablePorts {
     current_port: u16,
